fix: move to admin-ui package (ui-new)
