+++ packages/admin-ui/tsconfig.json
@@ -1,14 +1,34 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src", "__tests__"],
-  "references": [],
+  "include": [
+    "src",
+    "__tests__"
+  ],
+  "references": [
+    {
+      "path": "../react-composition"
+    }
+  ],
   "compilerOptions": {
-    "rootDirs": ["./src", "./__tests__"],
+    "rootDirs": [
+      "./src",
+      "./__tests__"
+    ],
     "outDir": "./dist",
     "declarationDir": "./dist",
     "paths": {
-      "~/*": ["./src/*"],
-      "~tests/*": ["./__tests__/*"]
+      "~/*": [
+        "./src/*"
+      ],
+      "~tests/*": [
+        "./__tests__/*"
+      ],
+      "@webiny/react-composition/*": [
+        "../react-composition/src/*"
+      ],
+      "@webiny/react-composition": [
+        "../react-composition/src"
+      ]
     },
     "baseUrl": "."
   }
