--- packages/ui-new/src/Progress/index.ts
@@ -1,1 +1,0 @@
-export * from "./CircularProgress";
